1. Refactor get_candidates, get_candidates_from_text, resolve_system_entity, parse_numerics into DucklingRecognizer 2. Update QueryFactory to retrieve the SystemEntityRecognizer from app_path 3. Update methods to pass in sys_recognizer / sys_resolver 4. Update tests
