+++ mindmeld/models/taggers/taggers.py
@@ -25,7 +25,7 @@ from ...core import (
     _sort_by_lowest_time_grain,
 )
 from ...markup import MarkupError
-from ...ser import SystemEntityResolutionError, resolve_system_entity
+from ...system_entity_recognizer import SystemEntityResolutionError, DucklingRecognizer
 from ..helpers import ENABLE_STEMMING, get_feature_extractor
 
 logger = logging.getLogger(__name__)
@@ -260,7 +260,7 @@ def _get_tags_from_entities(query, entities, scheme="IOB"):
     return iobs, types
 
 
-def get_entities_from_tags(query, tags, duckling_url=None):
+def get_entities_from_tags(query, tags, sys_resolver=None):
     """From a set of joint IOB tags, parse the app and system entities.
 
     This performs the reverse operation of get_tags_from_entities.
@@ -269,7 +269,7 @@ def get_entities_from_tags(query, tags, duckling_url=None):
         query (Query): Any query instance.
         tags (list of str): Joint app and system tags, like those
             created by get_tags_from_entities.
-        duckling_url (str): The URL for Duckling, if not passed, use the default one.
+        sys_resolver (SystemEntityRecognizer): Default to Duckling
 
     Returns:
         (list of QueryEntity) The tuple containing the list of entities.
@@ -277,6 +277,8 @@ def get_entities_from_tags(query, tags, duckling_url=None):
     normalized_tokens = query.normalized_tokens
 
     entities = []
+
+    sys_resolver = sys_resolver or DucklingRecognizer.get_instance()
 
     def _is_system_entity(entity_type):
         if entity_type.split("_")[0] == "sys":
@@ -309,7 +311,7 @@ def get_entities_from_tags(query, tags, duckling_url=None):
         span = query.transform_span(norm_span, TEXT_FORM_NORMALIZED, TEXT_FORM_RAW)
 
         try:
-            entity = resolve_system_entity(query, entity_type, span, url=duckling_url)
+            entity = sys_resolver.resolve_system_entity(query, entity_type, span)
             entities.append(entity)
             logger.debug("Appended system entity %s.", entity)
         except SystemEntityResolutionError:
